requirements.txt
fastapi
uvicorn
openai
sqlalchemy
pgvector
psycopg2-binary
pydantic
httpx
python-dotenv
alembic